WiFiManager.cpp: replaced root with wifi scan page
This patch replaces the root page with the wifi scan page, immediately
showing the list of networks upon connecting to the ESP AP.

The old root page only contained a link to the wifi scan page and
a link to the wifi page for direct input. Since the wifi scan page also
provided direct input, there is no need for a separate root page or
separate wifi page for direct input.